stomatology/requirements.txt
Django~=4.2.20
django-phonenumber-field==8.0.0
phonenumberslite==8.12.11
psycopg2-binary==2.9.10

gevent==24.11.1
gunicorn==23.0.0